+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 @ 2025-06-02, v1.0.61
 
- * Added simple key-value tests
+ * Added simple key-value tests and xcache
  * Test selected-run by pattern now not exact name
 
 @ 2025-05-30, v1.0.60